# go/server/openapi/model_book.go
/*
 * Books API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 0.0.1
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type Book struct {

	Id int32 `json:"id,omitempty"`

	Title string `json:"title"`

	Author string `json:"author"`
}